+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Changed
+- Asset Request details
+- Withdraw Request details
+
 ## [1.14.0-rc.0] - 2021-02-09
 #### Added
 - Ability to delete key-value